+++ packages/start-static-server-functions/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @tanstack/start-static-server-functions
 
+## 1.167.17
+
+### Patch Changes
+
+- Updated dependencies [[`9bebf8d`](https://github.com/TanStack/router/commit/9bebf8dc9f2bf74b680c065a5aa63d03b9622825)]:
+  - @tanstack/react-start@1.168.25
+  - @tanstack/solid-start@1.168.25
+  - @tanstack/start-client-core@1.170.12
+
+## 1.167.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.170.11
+  - @tanstack/react-start@1.168.24
+  - @tanstack/solid-start@1.168.24
+
+## 1.167.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.170.10
+  - @tanstack/react-start@1.168.23
+  - @tanstack/solid-start@1.168.23
+
+## 1.167.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.170.9
+  - @tanstack/react-start@1.168.22
+  - @tanstack/solid-start@1.168.22
+
+## 1.167.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-start@1.168.21
+  - @tanstack/solid-start@1.168.21
+  - @tanstack/start-client-core@1.170.8
+
 ## 1.167.12
 
 ### Patch Changes